screencast: Remove swap_buffer function
Dequeuing a new buffer imidiately creates problem when the buffer is
destroid while renegotiation a working modifier because we end up using
a buffer which was freed while waiting for the buffer_done event. Then
xdpw will segfault when requesting a copy with a nonexisting buffer. To fix
this we can just dequeue a buffer right before we need it. This makes
the fallback via on_process obsolete.