screencast: stop assuming that allocation with implicit modifier works
It has shown that the assumption: "Allocation with implicit modifier will
always be available" doesn't hold true in all cases. Thus if allocation
with modifiers fails or is not available we will try to allocate a
buffer with implicit modifiers. If this allocation fails we mark the
session to avoid dmabufs, thus falling back to shm transport.